move to marker utils
